feat: переносит все функции для создания массива с описаниями фото в отдельный модуль
